# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Atlas Backend API",
        "description": "API documentation for Atlas Backend - Project Management System",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/pg/academic-periods": {
            "get": {
                "tags": [
                    "Academic Periods"
                ],
                "summary": "Get all academic periods",
                "operationId": "f0d7c0cc8f0c91e548e25d41d815d55d",
                "responses": {
                    "200": {
                        "description": "List of academic periods with their state name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AcademicPeriodSummary"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Academic Periods"
                ],
                "summary": "Create a new academic period",
                "operationId": "0473bc1f48a8c7e2dba06a19c4a46ca1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2025-1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Academic period created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicPeriodResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/pg/academic-periods/{id}": {
            "get": {
                "tags": [
                    "Academic Periods"
                ],
                "summary": "Get a specific academic period",
                "operationId": "90d091cef8f748cf35467b88c5bce541",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic period details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicPeriodResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Academic period not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Academic Periods"
                ],
                "summary": "Update an academic period",
                "operationId": "e096c1c3d099f5a8477c9c8618d6df65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2025-1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Academic period updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicPeriodResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Academic period not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Academic Periods"
                ],
                "summary": "Delete an academic period",
                "operationId": "da734672bab8964ba9c5915f8fb9b46b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic period deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Academic period deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Academic period not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AcademicPeriodSummary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "2025-1"
                    },
                    "state_name": {
                        "type": "string",
                        "example": "Activo",
                        "nullable": true
                    },
                    "state_description": {
                        "type": "string",
                        "example": "Periodo en curso",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AcademicPeriodResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "2025-1"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-30"
                    },
                    "state_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "state_name": {
                        "type": "string",
                        "example": "Activo",
                        "nullable": true
                    },
                    "state_description": {
                        "type": "string",
                        "example": "Periodo en curso",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T18:30:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}